Ajout fin des Mappers

+++ src/main/java/com/forge_miniatures/dto/ContainIdDTO.java
@@ -8,6 +8,6 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 public class ContainIdDTO {
-    private int idtPanier;
-    private int idtArticle;
+    private Long idtPanier;
+    private Long idtArticle;
 }
